+++ spring-oxygen-flysql/src/main/java/com/isxcode/oxygen/flysql/core/Flysql.java
@@ -2,7 +2,9 @@ package com.isxcode.oxygen.flysql.core;
 
 import com.isxcode.oxygen.flysql.constant.FlysqlConstants;
 import com.isxcode.oxygen.flysql.entity.FlysqlKey;
+import com.isxcode.oxygen.flysql.enums.DataBaseType;
 import com.isxcode.oxygen.flysql.enums.SqlType;
+import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.jdbc.core.JdbcTemplate;
 
 import javax.sql.DataSource;
@@ -19,11 +21,15 @@ public class Flysql {
     /**
      * store datasource list
      */
-    private static Map<String, JdbcTemplate> jdbcTemplateMap;
+    private final Map<String, JdbcTemplate> jdbcTemplateMap;
 
-    public Flysql(Map<String, JdbcTemplate> jdbcTemplateMap) {
+    private final Map<String, MongoTemplate> mongdTemplateMap;
 
-        Flysql.jdbcTemplateMap = jdbcTemplateMap;
+    public Flysql(Map<String, JdbcTemplate> jdbcTemplateMap,
+                  Map<String, MongoTemplate> mongdTemplateMap) {
+
+        this.mongdTemplateMap = mongdTemplateMap;
+        this.jdbcTemplateMap = jdbcTemplateMap;
     }
 
     /**
@@ -31,7 +37,7 @@ public class Flysql {
      *
      * @return DataSource
      */
-    public static DataSource getDefaultDataSource() {
+    public DataSource getDefaultDataSource() {
 
         return jdbcTemplateMap.get(FlysqlConstants.PRIMARY_DATASOURCE_NAME).getDataSource();
     }
@@ -45,9 +51,9 @@ public class Flysql {
      * @return FlysqlBuilder
      * @since 0.0.1
      */
-    public static <A> FlysqlBuilder<A> insert(String dataSourceName, Class<A> targetClass) {
+    public <A> FlysqlBuilder<A> insert(String dataSourceName, Class<A> targetClass) {
 
-        return new FlysqlBuilder<>(new FlysqlKey<>(SqlType.INSERT, jdbcTemplateMap.get(dataSourceName), targetClass));
+        return new FlysqlBuilder<>(new FlysqlKey<>(DataBaseType.MYSQL, SqlType.INSERT, jdbcTemplateMap.get(dataSourceName), targetClass));
     }
 
     /**
@@ -58,7 +64,7 @@ public class Flysql {
      * @return FlysqlBuilder
      * @since 0.0.1
      */
-    public static <A> FlysqlBuilder<A> insert(Class<A> targetClass) {
+    public <A> FlysqlBuilder<A> insert(Class<A> targetClass) {
 
         return insert(FlysqlConstants.PRIMARY_DATASOURCE_NAME, targetClass);
     }
@@ -72,9 +78,9 @@ public class Flysql {
      * @return FlysqlBuilder
      * @since 0.0.1
      */
-    public static <A> FlysqlBuilder<A> delete(String dataSourceName, Class<A> targetClass) {
+    public <A> FlysqlBuilder<A> delete(String dataSourceName, Class<A> targetClass) {
 
-        return new FlysqlBuilder<>(new FlysqlKey<>(SqlType.DELETE, jdbcTemplateMap.get(dataSourceName), targetClass));
+        return new FlysqlBuilder<>(new FlysqlKey<>(DataBaseType.MYSQL, SqlType.DELETE, jdbcTemplateMap.get(dataSourceName), targetClass));
     }
 
     /**
@@ -85,7 +91,7 @@ public class Flysql {
      * @return FlysqlBuilder
      * @since 0.0.1
      */
-    public static <A> FlysqlBuilder<A> delete(Class<A> targetClass) {
+    public <A> FlysqlBuilder<A> delete(Class<A> targetClass) {
 
         return delete(FlysqlConstants.PRIMARY_DATASOURCE_NAME, targetClass);
     }
@@ -99,9 +105,9 @@ public class Flysql {
      * @return FlysqlBuilder
      * @since 0.0.1
      */
-    public static <A> FlysqlBuilder<A> update(String dataSourceName, Class<A> targetClass) {
+    public <A> FlysqlBuilder<A> update(String dataSourceName, Class<A> targetClass) {
 
-        return new FlysqlBuilder<>(new FlysqlKey<>(SqlType.UPDATE, jdbcTemplateMap.get(dataSourceName), targetClass));
+        return new FlysqlBuilder<>(new FlysqlKey<>(DataBaseType.MYSQL, SqlType.UPDATE, jdbcTemplateMap.get(dataSourceName), targetClass));
     }
 
     /**
@@ -112,7 +118,7 @@ public class Flysql {
      * @return FlysqlBuilder
      * @since 0.0.1
      */
-    public static <A> FlysqlBuilder<A> update(Class<A> targetClass) {
+    public <A> FlysqlBuilder<A> update(Class<A> targetClass) {
 
         return update(FlysqlConstants.PRIMARY_DATASOURCE_NAME, targetClass);
     }
@@ -127,9 +133,9 @@ public class Flysql {
      * @return FlysqlBuilder
      * @since 0.0.1
      */
-    public static <A> FlysqlBuilder<A> view(String dataSourceName, String viewName, Class<A> targetClass) {
+    public <A> FlysqlBuilder<A> view(String dataSourceName, String viewName, Class<A> targetClass) {
 
-        return new FlysqlBuilder<>(new FlysqlKey<>(SqlType.VIEW, jdbcTemplateMap.get(dataSourceName), targetClass, viewName));
+        return new FlysqlBuilder<>(new FlysqlKey<>(DataBaseType.MYSQL, SqlType.VIEW, jdbcTemplateMap.get(dataSourceName), targetClass, viewName));
     }
 
     /**
@@ -141,7 +147,7 @@ public class Flysql {
      * @return FlysqlBuilder
      * @since 0.0.1
      */
-    public static <A> FlysqlBuilder<A> view(String viewName, Class<A> targetClass) {
+    public <A> FlysqlBuilder<A> view(String viewName, Class<A> targetClass) {
 
         return view(FlysqlConstants.PRIMARY_DATASOURCE_NAME, viewName, targetClass);
     }
@@ -155,9 +161,9 @@ public class Flysql {
      * @return FlysqlBuilder
      * @since 0.0.1
      */
-    public static <A> FlysqlBuilder<A> select(String dataSourceName, Class<A> targetClass) {
+    public <A> FlysqlBuilder<A> select(String dataSourceName, Class<A> targetClass) {
 
-        return new FlysqlBuilder<>(new FlysqlKey<>(SqlType.SELECT, jdbcTemplateMap.get(dataSourceName), targetClass));
+        return new FlysqlBuilder<>(new FlysqlKey<>(DataBaseType.MYSQL, SqlType.SELECT, jdbcTemplateMap.get(dataSourceName), targetClass));
     }
 
     /**
@@ -168,9 +174,146 @@ public class Flysql {
      * @return FlysqlBuilder
      * @since 0.0.1
      */
-    public static <A> FlysqlBuilder<A> select(Class<A> targetClass) {
+    public <A> FlysqlBuilder<A> select(Class<A> targetClass) {
 
         return select(FlysqlConstants.PRIMARY_DATASOURCE_NAME, targetClass);
+    }
+
+    /**
+     * insert builder
+     *
+     * @param <A>            A
+     * @param targetClass    targetClass
+     * @param dataSourceName dataSourceName
+     * @return FlysqlBuilder
+     * @since 0.0.1
+     */
+    public <A> FlysqlBuilder<A> mongoInsert(String dataSourceName, Class<A> targetClass) {
+
+        return new FlysqlBuilder<>(new FlysqlKey<>(DataBaseType.MONGO, SqlType.INSERT, mongdTemplateMap.get(dataSourceName), targetClass));
+    }
+
+    /**
+     * insert builder
+     *
+     * @param targetClass targetClass
+     * @param <A>         A
+     * @return FlysqlBuilder
+     * @since 0.0.1
+     */
+    public <A> FlysqlBuilder<A> mongoInsert(Class<A> targetClass) {
+
+        return mongoInsert(FlysqlConstants.PRIMARY_DATASOURCE_NAME, targetClass);
+    }
+
+    /**
+     * delete builder
+     *
+     * @param targetClass    targetClass
+     * @param dataSourceName dataSourceName
+     * @param <A>            A
+     * @return FlysqlBuilder
+     * @since 0.0.1
+     */
+    public <A> FlysqlBuilder<A> mongoDelete(String dataSourceName, Class<A> targetClass) {
+
+        return new FlysqlBuilder<>(new FlysqlKey<>(DataBaseType.MONGO, SqlType.DELETE, mongdTemplateMap.get(dataSourceName), targetClass));
+    }
+
+    /**
+     * delete builder
+     *
+     * @param targetClass targetClass
+     * @param <A>         A
+     * @return FlysqlBuilder
+     * @since 0.0.1
+     */
+    public <A> FlysqlBuilder<A> mongoDelete(Class<A> targetClass) {
+
+        return mongoDelete(FlysqlConstants.PRIMARY_DATASOURCE_NAME, targetClass);
+    }
+
+    /**
+     * update builder
+     *
+     * @param targetClass    targetClass
+     * @param dataSourceName dataSourceName
+     * @param <A>            A
+     * @return FlysqlBuilder
+     * @since 0.0.1
+     */
+    public <A> FlysqlBuilder<A> mongoUpdate(String dataSourceName, Class<A> targetClass) {
+
+        return new FlysqlBuilder<>(new FlysqlKey<>(DataBaseType.MONGO, SqlType.UPDATE, mongdTemplateMap.get(dataSourceName), targetClass));
+    }
+
+    /**
+     * update builder
+     *
+     * @param targetClass targetClass
+     * @param <A>         A
+     * @return FlysqlBuilder
+     * @since 0.0.1
+     */
+    public <A> FlysqlBuilder<A> mongoUpdate(Class<A> targetClass) {
+
+        return mongoUpdate(FlysqlConstants.PRIMARY_DATASOURCE_NAME, targetClass);
+    }
+
+    /**
+     * view builder
+     *
+     * @param targetClass    targetClass
+     * @param dataSourceName dataSourceName
+     * @param viewName       viewName
+     * @param <A>            A
+     * @return FlysqlBuilder
+     * @since 0.0.1
+     */
+    public <A> FlysqlBuilder<A> mongoView(String dataSourceName, String viewName, Class<A> targetClass) {
+
+        return new FlysqlBuilder<>(new FlysqlKey<>(DataBaseType.MONGO, SqlType.VIEW, mongdTemplateMap.get(dataSourceName), targetClass, viewName));
+    }
+
+    /**
+     * view builder
+     *
+     * @param targetClass targetClass
+     * @param viewName    viewName
+     * @param <A>         A
+     * @return FlysqlBuilder
+     * @since 0.0.1
+     */
+    public <A> FlysqlBuilder<A> mongoView(String viewName, Class<A> targetClass) {
+
+        return mongoView(FlysqlConstants.PRIMARY_DATASOURCE_NAME, viewName, targetClass);
+    }
+
+    /**
+     * select builder
+     *
+     * @param targetClass    targetClass
+     * @param dataSourceName dataSourceName
+     * @param <A>            A
+     * @return FlysqlBuilder
+     * @since 0.0.1
+     */
+    public <A> FlysqlBuilder<A> mongoSelect(String dataSourceName, Class<A> targetClass) {
+
+        return new FlysqlBuilder<>(new FlysqlKey<>(DataBaseType.MONGO, SqlType.SELECT, mongdTemplateMap.get(dataSourceName), targetClass));
+    }
+
+    /**
+     * select builder
+     *
+     * @param targetClass targetClass
+     * @param <A>         A
+     * @return FlysqlBuilder
+     * @since 0.0.1
+     */
+    public <A> FlysqlBuilder<A> mongoSelect(Class<A> targetClass) {
+
+        return mongoSelect(FlysqlConstants.PRIMARY_DATASOURCE_NAME, targetClass);
     }
 
 }