+++ spring-oxygen-flysql/src/main/java/com/isxcode/oxygen/flysql/properties/FlysqlDataSourceProperties.java
@@ -2,6 +2,7 @@ package com.isxcode.oxygen.flysql.properties;
 
 import lombok.Data;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceProperties;
+import org.springframework.boot.autoconfigure.mongo.MongoProperties;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 import java.util.Map;
@@ -17,7 +18,13 @@ import java.util.Map;
 public class FlysqlDataSourceProperties {
 
     /**
-     * datasource config
+     *
      */
     private Map<String, DataSourceProperties> datasource;
+
+    /**
+     * 配置mongodb
+     */
+    private Map<String, MongoProperties> mongodb;
+
 }