+++ spring-oxygen-flysql/src/main/java/com/isxcode/oxygen/flysql/utils/FlysqlUtils.java
@@ -21,6 +21,7 @@ public class FlysqlUtils {
 
     /**
      * parse class to map
+     * 记录对象，各个字段的映射关系
      *
      * @param genericType targetClass
      * @return Map[propertiesName, type]
@@ -28,19 +29,28 @@ public class FlysqlUtils {
      */
     public static Map<String, String> parseBeanProperties(Class<?> genericType) {
 
+        // 使用spring自带工具进行解析
         PropertyDescriptor[] propertyDescriptors = BeanUtils.getPropertyDescriptors(genericType);
+
+        // 返回结果
         Map<String, String> columnNameMap = new HashMap<>(propertyDescriptors.length - 1);
 
         for (PropertyDescriptor propertyMeta : propertyDescriptors) {
+
+            // 排除class类型
             if (!ReflectConstants.CLASS.equals(propertyMeta.getName())) {
+
                 try {
                     Field metaField = propertyMeta.getReadMethod().getDeclaringClass().getDeclaredField(propertyMeta.getName());
+
+                    // 如果用户自定义注解，则使用用户自己定义的字段
                     if (metaField.isAnnotationPresent(ColumnName.class)) {
                         columnNameMap.put(propertyMeta.getName(), metaField.getAnnotation(ColumnName.class).value());
                     } else {
                         columnNameMap.put(propertyMeta.getName(), ReflectUtils.humpToLine(propertyMeta.getName()));
                     }
                 } catch (NoSuchFieldException e) {
+                    // 如果映射失败，则默认使用下划线分割字段名
                     columnNameMap.put(propertyMeta.getName(), ReflectUtils.humpToLine(propertyMeta.getName()));
                 }
             }