+++ CHANGELOG.md
@@ -1,14 +1,35 @@
 ## 0.0.2
 
-### ✨ Feature
+### 💥️ 重大变动
 
-- add new features oxygen-cli
-- support new datasource -- Oracle
+- Flysql从静态导入转为注入方式
+- Flysql使用语法变法
 
-### 🎨 Enhancement
+```java
+class demo{
+    
+    private final Flysql flysql;
+    
+    public demo(Flysql flysql){
+        this.flysql = flysql;
+    }
+    
+    void test(){
+        flysql.build().select().eq().query();    
+    }   
+}
+```
 
-- enhance key column update
+### ✨ 新功能
 
-### 🐛 Fix
+- 支持新的数据源 -- Oracle
+- 支持新的数据源 -- MongoDB
 
-- fix can not create default user
+### 🎨 优化
+
+- 优化打包大小，提出lombok以来
+- 添加系统性测试
+
+### 🐛 修复
+
+-